+++ public/modules/areas.js
@@ -9,87 +9,51 @@ let areaColors = { //Ithai adjustments Nov 24
   "ACTING": "#fb4b4b", //
   "MOVEMENT": "#ffd08f", //
   "SOUND": "#2ec4b6", //
-  "TECHNOLOGY": "#9bff94", //
+  "TECH": "#9bff94", //
   "WRITING": "#7aa7ff", //
   "VISUAL ART": "#d90368", // 
   "STUDIES": "#fcff52", //
 };
 
-// let areaColors = { //my adjustments 4-19
-//   "CORE": "#f72585", //
-//   "SOUL": "#eeeeff", //
-//   "IMAGE": "#cd20ff", //
-//   "ACTING": "#d64933", //
-//   "MOVEMENT": "#ff7900", //
-//   "SOUND": "#2ec4b6", //
-//   "TECHNOLOGY": "#38b000", //
-//   "WRITING": "#2D99C8", //
-//   "VISUAL ART": "#d90368", // 
-//   "STUDIES": "#ffd407", //
-// };
-
-// let areaColors = { //despina 4/18
-//   "CORE": "#f72585", //
-//   "SOUL": "#eeeeff", //
-//   "IMAGE": "#2ec4b6", //
-//   "ACTING": "#d64933", //
-//   "MOVEMENT": "#ff7900", //
-//   "SOUND": "#cd20ff", //
-//   "TECHNOLOGY": "#d90368", //
-//   "WRITING": "#3a86ff", //
-//   "VISUAL ART": "#38b000", // 
-//   "STUDIES": "#2b50aa", //
-// };
-
 let areas = [
-  ["CORE", areaColors["CORE"]], //thulian pink (palevioletred css)
-  ["SOUL", areaColors["SOUL"]], //whiteblue
-  ["IMAGE", areaColors["IMAGE"]], //aero blue
-  ["ACTING", areaColors["ACTING"]], //goldenrod
-  ["MOVEMENT", areaColors["MOVEMENT"]], //pomp and power (dark lilac)
-  ["SOUND", areaColors["SOUND"]], //tangerine
-  ["TECHNOLOGY", areaColors["TECHNOLOGY"]], //forest green
-  ["VISUAL ART", areaColors["VISUAL ART"]], // imperial red
-  ["STUDIES", areaColors["STUDIES"]], //pale dogwood (pink)
-  ["WRITING", areaColors["WRITING"]], //celestial blue
+  ["CORE", areaColors["CORE"]], 
+  ["SOUL", areaColors["SOUL"]], 
+  ["IMAGE", areaColors["IMAGE"]], 
+  ["ACTING", areaColors["ACTING"]], 
+  ["MOVEMENT", areaColors["MOVEMENT"]], 
+  ["SOUND", areaColors["SOUND"]], 
+  ["TECH", areaColors["TECH"]], 
+  ["VISUAL ART", areaColors["VISUAL ART"]], 
+  ["STUDIES", areaColors["STUDIES"]], 
+  ["WRITING", areaColors["WRITING"]], 
 ];
 
-let mainKeywords = [ //main 24
-"acting",
-"art studio",
-"coding",
-"collaboration",
-"critical making",
-"critical thinking",
-"design",
-"directing",
-"electronics",
-"fabrication",
-"fashion/wearable",
-"film/video",
-"guest artists",
-"interactive",
-"lighting",
-"live nyc",
-"live performance",
-"movement",
-"music/sound",
-"producing",
-"research",
-"theater",
-"visual art",
-"writing",
-]
-
-// let areaColors = {//silly, shh
-//   "CORE": "#DB7093", //thulian pink (palevioletred css)
-//   "SOUL": "#eeeeff", //whiteblue
-//   "IMAGE": "#14BDEB", //aero blue
-//   "ACTING": "#f5dc23", //goldenrod
-//   "MOVEMENT": "#89608E", //pomp and power (dark lilac)
-//   "SOUND": "#F08700", //tangerine
-//   "TECHNOLOGY": "#428722", //forest green
-//   "WRITING": "#5792C3", //celestial blue
-//   "VISUAL ART": "#FB3640", // imperial red
-//   "STUDIES": "#fac9b8", //pale dogwood (pink)
-// };
+let mainKeywords = [ //27 as of Feb '25
+  "acting",
+  "animation",
+  "art studio",
+  "coding",
+  "collaboration",
+  "computer graphics/3D",
+  "computer software",
+  "critical making",
+  "directing/producing",
+  "electronics/hardware",
+  "environment/climate",
+  "fabrication/sculpture",
+  "fashion/wearables",
+  "film/video",
+  "interactive/games",
+  "lighting",
+  "live performance",
+  "movement/dance",
+  "music/sound",
+  "outings",
+  "photography",
+  "printed matter/graphic design",
+  "process/iteration",
+  "research and discussion",
+  "theater",
+  "visual art",
+  "writing"
+];